create views

+++ src/main/java/br/edu/ifg/projetopraticoweb/controller/TaskController.java
@@ -4,6 +4,7 @@ import br.edu.ifg.projetopraticoweb.dto.TaskDTO;
 import br.edu.ifg.projetopraticoweb.exception.ResourceNotFoundException;
 import br.edu.ifg.projetopraticoweb.mapper.TaskMapper;
 import br.edu.ifg.projetopraticoweb.model.Task;
+import br.edu.ifg.projetopraticoweb.service.AuthenticationService;
 import br.edu.ifg.projetopraticoweb.service.TaskService;
 import br.edu.ifg.projetopraticoweb.validator.TaskDTOValidator;
 import org.springframework.http.HttpStatus;
@@ -13,6 +14,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
@@ -22,10 +24,12 @@ public class TaskController {
     private final TaskService taskService;
     private final TaskMapper taskMapper;
     private final TaskDTOValidator validator;
+    private final AuthenticationService authenticationService;
 
-    public TaskController(TaskService taskService, TaskMapper taskMapper) {
+    public TaskController(TaskService taskService, TaskMapper taskMapper, AuthenticationService authenticationService) {
         this.taskService = taskService;
         this.taskMapper = taskMapper;
+        this.authenticationService = authenticationService;
         this.validator = new TaskDTOValidator();
     }
 
@@ -47,7 +51,7 @@ public class TaskController {
     }
 
     // Criar uma nova tarefa
-    @PostMapping
+    @PostMapping("/new")
     public ResponseEntity<String> createTask(@RequestBody TaskDTO taskDTO) {
         if (!validator.isValid(taskDTO)) {
             return ResponseEntity.badRequest().body("Dados inválidos");
@@ -55,6 +59,7 @@ public class TaskController {
 
         try {
             Task task = taskMapper.toEntity(taskDTO);
+            task.setUser(authenticationService.getAuthenticatedUser());
             taskService.save(task);
             return ResponseEntity.created(URI.create("/tasks/" + task.getId())).build();
         } catch (ResourceNotFoundException e) {
@@ -65,6 +70,13 @@ public class TaskController {
     // Atualizar uma tarefa existente
     @PutMapping("/{id}")
     public ResponseEntity<String> updateTask(@PathVariable Long id, @RequestBody TaskDTO taskDTO) {
+        Optional<Task> taskOptional = taskService.findById(id);
+        if (!authenticationService.authenticate(taskOptional.orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Task não encontrada")
+        ).getUser().getEmail())) {
+            return ResponseEntity.badRequest().body("Usuário não autorizado");
+        }
+
         if (!validator.isValid(taskDTO)) {
             return ResponseEntity.badRequest().body("Dados inválidos");
         }
@@ -72,6 +84,7 @@ public class TaskController {
         try {
             Task task = taskMapper.toEntity(taskDTO);
             task.setId(id);
+            task.setUser(taskOptional.get().getUser());
             taskService.save(task);
             return ResponseEntity.ok().build();
         } catch (ResourceNotFoundException e) {
@@ -82,6 +95,13 @@ public class TaskController {
     // Deletar uma tarefa por ID
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteTask(@PathVariable Long id) {
+        Optional<Task> taskOptional = taskService.findById(id);
+        if (!authenticationService.authenticate(taskOptional.orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Task não encontrada")
+        ).getUser().getEmail())) {
+            return ResponseEntity.badRequest().body("Usuário não autorizado");
+        }
+
         try {
             taskService.delete(id);
             return ResponseEntity.noContent().build();
